Restructured the exception throwing mechanism for better consistency
There is a new class, HTTPException, which are explicitly catched
by ServerMain - ServerMain is now in charge of setting the error
code in the HTTP response and printing the template. Before, it
was handled directly in the exceptions themselves, but it made the
message string ambiguous - sometimes it was a text message, sometimes
a rendered HTML template. Now, messages are always in text form, but
when an Exception extends HTTPException, it has a template attribute
that ServerMain uses to render the exception.