Abstracted the URLMapping and Controller classes into interfaces
URLMapping is now a interface capable of resolving an URL.
This makes redefining how the framework maps an URL to a controller
by having the URL matching and parameter extraction customisable.
By default, the default Module class makes use of the
ERegURLMapping implementation (which was there before by default as
the URLMapping class)

The Controller class is now an interface, providing a required
interface for controller operation, should someone want to implement
a different mechanism. By default, the AbstractController can be used.